requirements.txt
adafruit-io==2.8.0
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.4
bcrypt==4.0.1
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cryptography==44.0.2
cycler==0.12.1
ecdsa==0.19.0
exceptiongroup==1.1.3
fastapi==0.104.1
fonttools==4.45.1
h11==0.14.0
idna==3.6
importlib-resources==6.1.1
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
paho-mqtt==2.1.0
pandas==2.1.3
passlib==1.7.4
Pillow==10.1.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
schedule==1.2.1
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tabulate==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.3.0
uvicorn==0.24.0
zipp==3.17.0
